+++ design-at-cornell/src/explore-courses/courses/Sort.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import {
+  Divider, 
+  SortBy
+} from '../ExploreCoursesStyles';
+
+const Sort = () => (
+  <Divider>
+    <SortBy>
+      <p>Sort By</p>
+      <img src={require('../../static/images/sort-triangle.svg')} alt="sort" />
+    </SortBy>
+  </Divider>
+)
+
+export default Sort;
